backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
mongodb==1.0.0
motor==3.5.1
python-dotenv==1.0.1
passlib==1.7.4
bcrypt==4.2.0
python-jose==3.3.0
pydantic==2.9.0
pydantic-settings==2.5.2
